requirements.txt
alembic==1.9.4
click==8.0.3
dnspython==2.2.0
email-validator==1.3.1
Flask==2.2.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
python-dotenv==0.21.1
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.13.0
